Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -3,7 +3,7 @@
 int main(){
     float n;
     scanf("%f",&n);
-    float ans=round(n*100.0)/100;
+    double ans=round(n*100.0)/100;
 
     printf("You entered: %f",ans);
     return 0;
